% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml.R
\name{asis_yaml_output}
\alias{asis_yaml_output}
\title{Export yml object as a YAML knitr code chunk}
\usage{
asis_yaml_output(.yml, fences = TRUE)
}
\arguments{
\item{.yml}{a \code{yml} object created by \code{yml()}, \code{as_yml()}, or returned by
a \code{yml_*()} function}

\item{fences}{Logical. Write fences ("---") before and after YAML?}
}
\description{
\code{asis_yaml_output()} exports a \code{yml} object as a YAML knitr code chunk
instead of as an R object. Doing so adds code highlighting for YAML syntax.
}
\seealso{
Other yml: \code{\link{bib2yml}},
  \code{\link{draw_yml_tree}}, \code{\link{has_field}},
  \code{\link{read_json}}, \code{\link{use_yml_defaults}},
  \code{\link{use_yml_file}}, \code{\link{use_yml}},
  \code{\link{yml_author}},
  \code{\link{yml_blogdown_opts}},
  \code{\link{yml_bookdown_opts}},
  \code{\link{yml_citations}}, \code{\link{yml_clean}},
  \code{\link{yml_distill_opts}},
  \code{\link{yml_latex_opts}}, \code{\link{yml_output}},
  \code{\link{yml_pagedown_opts}},
  \code{\link{yml_params}}, \code{\link{yml_pkgdown}},
  \code{\link{yml_reference}}, \code{\link{yml_replace}},
  \code{\link{yml_resource_files}},
  \code{\link{yml_rsconnect_email}},
  \code{\link{yml_rticles_opts}},
  \code{\link{yml_runtime}}, \code{\link{yml_site_opts}},
  \code{\link{yml_toc}}, \code{\link{yml_vignette}}
}
\concept{yml}
