% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_bookdown.R
\name{gitbook_config}
\alias{gitbook_config}
\title{Title}
\usage{
gitbook_config(toc_collapse = yml_blank(),
  toc_scroll_highlight = yml_blank(), toc_before = yml_blank(),
  toc_after = yml_blank(), toolbar_position = yml_blank(),
  edit = yml_blank(), download = yml_blank(), search = yml_blank(),
  fontsettings_theme = yml_blank(), fontsettings_family = yml_blank(),
  fontsettings_size = yml_blank(), sharing_facebook = yml_blank(),
  sharing_twitter = yml_blank(), sharing_google = yml_blank(),
  sharing_linkedin = yml_blank(), sharing_weibo = yml_blank(),
  sharing_instapaper = yml_blank(), sharing_vk = yml_blank(),
  sharing_all = yml_blank())
}
\arguments{
\item{sharing_all}{}
}
\description{
\itemize{
\item defaults and descriptions are available at https://bookdown.org/yihui/bookdown/html.html
}
}
