% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_rmarkdown.R
\name{shiny_date}
\alias{shiny_date}
\title{Title}
\usage{
shiny_date(label, value = NULL, min = NULL, max = NULL,
  format = "yyyy-mm-dd", startview = "month", weekstart = 0,
  language = "en", width = NULL, autoclose = TRUE,
  datesdisabled = NULL, daysofweekdisabled = NULL)
}
\arguments{
\item{daysofweekdisabled}{}
}
\description{
Title
}
