% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_rmarkdown.R
\name{shiny_select}
\alias{shiny_select}
\title{Title}
\usage{
shiny_select(label, choices, selected = NULL, multiple = FALSE,
  selectize = TRUE, width = NULL, size = NULL)
}
\arguments{
\item{size}{}
}
\description{
Title
}
