% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_rmarkdown.R
\name{shiny_slider}
\alias{shiny_slider}
\title{Title}
\usage{
shiny_slider(label, min, max, value, step = NULL, round = FALSE,
  format = NULL, locale = NULL, ticks = TRUE, animate = FALSE,
  width = NULL, sep = ",", pre = NULL, post = NULL,
  timeFormat = NULL, timezone = NULL, dragRange = TRUE)
}
\arguments{
\item{dragRange}{}
}
\description{
Title
}
