% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_citations.R
\name{yml_citations}
\alias{yml_citations}
\title{Title}
\usage{
yml_citations(bibliography = yml_blank(), csl = yml_blank(),
  citation_abbreviations = yml_blank(), natbib = yml_blank(),
  biblatex = yml_blank())
}
\arguments{
\item{biblatex}{}
}
\description{
Title
}
