% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml.R
\name{yml_handlers}
\alias{yml_handlers}
\title{Set handlers to process the way YAML is printed}
\usage{
yml_handlers()
}
\description{
ymlthis uses the yaml package to process and validate YAML; this package also
lets you specify how fields and values are printed using a list of handler
functions. \code{yml_handlers()} specifies defaults for the package used in the
print statement. See \code{\link[yaml:yaml.load]{yaml::yaml.load()}} for more on specifying handlers.
}
