% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_latex.R
\name{yml_latex_opts}
\alias{yml_latex_opts}
\title{Title}
\usage{
yml_latex_opts(.yml, block_headings = yml_blank(),
  classoption = yml_blank(), documentclass = yml_blank(),
  geometry = yml_blank(), indent = yml_blank(),
  linestretch = yml_blank(), margin_left = yml_blank(),
  margin_right = yml_blank(), margin_top = yml_blank(),
  margin_bottom = yml_blank(), pagestyle = yml_blank(),
  papersize = yml_blank(), secnumdepth = yml_blank(),
  fontenc = yml_blank(), fontfamily = yml_blank(),
  fontfamilyoptions = yml_blank(), fontsize = yml_blank(),
  mainfont = yml_blank(), sansfont = yml_blank(),
  monofont = yml_blank(), mathfont = yml_blank(),
  CJKmainfont = yml_blank(), mainfontoptions = yml_blank(),
  sansfontoptions = yml_blank(), monofontoptions = yml_blank(),
  mathfontoptions = yml_blank(), CJKoptions = yml_blank(),
  microtypeoptions = yml_blank(), colorlinks = yml_blank(),
  linkcolor = yml_blank(), filecolor = yml_blank(),
  citecolor = yml_blank(), urlcolor = yml_blank(),
  toccolor = yml_blank(), links_as_notes = yml_blank(),
  lof = yml_blank(), lot = yml_blank(), thanks = yml_blank(),
  toc = yml_blank(), toc_depth = yml_blank(),
  biblatexoptions = yml_blank(), biblio_style = yml_blank(),
  biblio_title = yml_blank(), bibliography = yml_blank(),
  natbiboptions = yml_blank())
}
\arguments{
\item{natbiboptions}{}
}
\description{
Title
}
