% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml.R
\name{yml_load}
\alias{yml_load}
\title{Load YAML from string}
\usage{
yml_load(x)
}
\arguments{
\item{x}{an object to pass to \code{\link[yaml:yaml.load]{yaml::yaml.load()}}}
}
\description{
\code{yml_load()} is a wrapper for \code{\link[yaml:yaml.load]{yaml::yaml.load()}} that also converts the
object to the \code{yml} class.
}
\examples{
c("title: my title", "author: Malcolm Barrett") \%>\%
  yml_load()

}
