% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_pkgdown.R
\name{yml_pkgdown}
\alias{yml_pkgdown}
\title{Title}
\usage{
yml_pkgdown(.yml, as_is = yml_blank(), extension = yml_blank())
}
\arguments{
\item{extension}{}
}
\description{
\itemize{
\item https://pkgdown.r-lib.org/articles/pkgdown.html
\item ?pkgdown::build_articles()
}
}
