% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_pkgdown.R
\name{yml_pkgdown_articles}
\alias{yml_pkgdown_articles}
\alias{pkgdown_article}
\title{Title}
\usage{
yml_pkgdown_articles(.yml, ...)

pkgdown_article(title = yml_blank(), desc = yml_blank(),
  contents = yml_blank())
}
\arguments{
\item{exclude}{}
}
\description{
\itemize{
\item https://pkgdown.r-lib.org/articles/pkgdown.html
\item ?pkgdown::build_articles
}
}
