% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_pkgdown.R
\name{yml_pkgdown_figures}
\alias{yml_pkgdown_figures}
\title{Title}
\usage{
yml_pkgdown_figures(.yml, figures = yml_blank(), dev = yml_blank(),
  dpi = yml_blank(), dev.args = yml_blank(), fig.ext = yml_blank(),
  fig.width = yml_blank(), fig.height = yml_blank(),
  fig.retina = yml_blank(), fig.asp = yml_blank())
}
\arguments{
\item{fig.asp}{}
}
\description{
\itemize{
\item ?pkgdown::build_reference
}
}
