% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_pkgdown.R
\name{yml_pkgdown_news}
\alias{yml_pkgdown_news}
\title{Title}
\usage{
yml_pkgdown_news(.yml, one_page = yml_blank())
}
\arguments{
\item{one_page}{}
}
\description{
\itemize{
\item https://pkgdown.r-lib.org/articles/pkgdown.html
}
}
