% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_replace.R
\name{yml_replace}
\alias{yml_replace}
\alias{yml_discard}
\title{Replace or discard top-level YAML fields}
\usage{
yml_replace(.yml, ...)

yml_discard(.yml, .rid)
}
\arguments{
\item{...}{}
}
\description{
Replace or discard top-level YAML fields
}
