% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_rticles.R
\name{yml_rticles_opts}
\alias{yml_rticles_opts}
\title{Title}
\usage{
yml_rticles_opts(.yml, title = yml_blank(), runninghead = yml_blank(),
  author = yml_blank(), authormark = yml_blank(),
  address = yml_blank(), corrauth = yml_blank(),
  corres = yml_blank(), email = yml_blank(), abstract = yml_blank(),
  received = yml_blank(), revised = yml_blank(),
  accepted = yml_blank(), keywords = yml_blank(),
  bibliography = yml_blank(), longtable = yml_blank(),
  classoption = yml_blank(), header_includes = yml_blank(),
  include_after = yml_blank())
}
\arguments{
\item{include_after}{}
}
\description{
\itemize{
\item ?rticles::acm_article
\item need to indicate which options belong to which functions
}
}
