% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_rmarkdown.R
\name{yml_site_opts}
\alias{yml_site_opts}
\title{Title}
\usage{
yml_site_opts(.yml, name = yml_blank(), favicon = yml_blank(),
  output_dir = yml_blank(), include = yml_blank(),
  exclude = yml_blank(), new_session = yml_blank())
}
\arguments{
\item{new_session}{}
}
\description{
\itemize{
\item https://bookdown.org/yihui/rmarkdown/rmarkdown-site.html#site_navigation
}
}
