% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_top.R
\name{yml_abstract}
\alias{yml_abstract}
\alias{yml_category}
\title{Title}
\usage{
yml_abstract(.yml, abstract)

yml_category(.yml, category)
}
\arguments{
\item{abstract}{}

\item{category}{}
}
\description{
Title

Title
}
