% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_citations.R
\name{bib2yml}
\alias{bib2yml}
\title{Convert bib files to YAML}
\usage{
bib2yml(.yml = NULL, path)
}
\arguments{
\item{.yml}{a \code{yml} object created by \code{yml()}, \code{as_yml()}, or returned by
a \code{yml_*()} function}

\item{path}{a path to the .bib file}
}
\value{
a \code{yml} object
}
\description{
\code{bib2yml()} uses pandoc to convert a .bib file to YAML. It also accepts an
optional \code{yml} object to prepend to the the YAML from the .bib file. If you
want to cite several R packages, see \code{\link[knitr:write_bib]{knitr::write_bib()}} to write a
bibliography file and convert it with \code{bib2yml()}.
}
\seealso{
Other yml: \code{\link{asis_yaml_output}},
  \code{\link{draw_yml_tree}},
  \code{\link{use_yml_defaults}},
  \code{\link{use_yml_file}}, \code{\link{use_yml}},
  \code{\link{yml_author}},
  \code{\link{yml_blogdown_opts}},
  \code{\link{yml_bookdown_opts}},
  \code{\link{yml_citations}}, \code{\link{yml_clean}},
  \code{\link{yml_distill_opts}},
  \code{\link{yml_latex_opts}}, \code{\link{yml_output}},
  \code{\link{yml_pagedown_opts}},
  \code{\link{yml_params}}, \code{\link{yml_pkgdown}},
  \code{\link{yml_reference}}, \code{\link{yml_replace}},
  \code{\link{yml_resource_files}},
  \code{\link{yml_rsconnect_email}},
  \code{\link{yml_rticles_opts}},
  \code{\link{yml_runtime}}, \code{\link{yml_site_opts}},
  \code{\link{yml_toc}}, \code{\link{yml_vignette}}

Other citations: \code{\link{yml_citations}},
  \code{\link{yml_reference}}
}
\concept{citations}
\concept{yml}
