% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_blogdown.R
\name{blogdown_template}
\alias{blogdown_template}
\alias{blogdown_archetypes}
\title{Create YAML based on blogdown theme archetypes}
\usage{
blogdown_template(type, path = ".", theme = NULL)

blogdown_archetypes(path = ".", theme = NULL)
}
\arguments{
\item{theme}{}
}
\description{
Create YAML based on blogdown theme archetypes
}
