% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_rmarkdown.R
\name{shiny_file}
\alias{shiny_file}
\title{Title}
\usage{
shiny_file(label, multiple = FALSE, accept = NULL, width = NULL,
  buttonLabel = "Browse...", placeholder = "No file selected")
}
\arguments{
\item{placeholder}{}
}
\description{
Title
}
