% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_rmarkdown.R
\name{shiny_radio}
\alias{shiny_radio}
\title{Title}
\usage{
shiny_radio(label, choices = NULL, selected = NULL, inline = FALSE,
  width = NULL, choiceNames = NULL, choiceValues = NULL)
}
\arguments{
\item{choiceValues}{}
}
\description{
Title
}
