% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_yml.R
\name{use_yml_file}
\alias{use_yml_file}
\alias{use_output_yml}
\alias{use_site_yml}
\alias{use_pkgdown_yml}
\alias{use_bookdown_yml}
\title{Write YAML to file}
\usage{
use_yml_file(x = NULL, path)

use_output_yml(x = NULL, path = ".")

use_site_yml(x = NULL, path = ".")

use_pkgdown_yml(x = NULL, path = ".")

use_bookdown_yml(x = NULL, path = ".")
}
\arguments{
\item{path}{}
}
\description{
Write YAML to file
}
