% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_helpers.R
\name{yml_code}
\alias{yml_code}
\title{Take code and write it as valid YAML}
\usage{
yml_code(x)
}
\arguments{
\item{x}{valid R code}
}
\value{
a character vector with class \code{verbatim}
}
\description{
\code{yml_code()} takes R code and writes it as valid YAML to be evaluated during
knitting. Note that \code{yml_code()} does not evaluate or validate the R code but
only captures it to use in the YAML field.
}
\examples{

yml_code(sys.Date())

}
\seealso{
\code{\link[=yml_verbatim]{yml_verbatim()}}
}
