% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_rmarkdown.R
\name{yml_navbar}
\alias{yml_navbar}
\title{Title}
\usage{
yml_navbar(.yml, title, type, left, right)
}
\arguments{
\item{right}{}
}
\description{
\itemize{
\item https://bookdown.org/yihui/rmarkdown/rmarkdown-site.html#site_navigation
\item ?rmarkdown::render_site
}
}
