% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_citations.R
\name{yml_reference}
\alias{yml_reference}
\alias{reference}
\title{Title}
\usage{
yml_reference(.yml, ..., .bibentry = NULL)

reference(id = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
See ?bibentry
}
